+++ dart/common/stl_allocator.hpp
@@ -92,6 +92,14 @@ public:
   /// @param[in] n: Number of objects earlier passed to allocate().
   void deallocate(pointer pointer, size_type n) noexcept;
 
+  /// Constructs an object at previously allocated storage.
+  template <typename U, typename... Args>
+  void construct(U* pointer, Args&&... args);
+
+  /// Destroys an object constructed by this allocator.
+  template <typename U>
+  void destroy(U* pointer) noexcept;
+
   /// Upper bound on elements that can be allocated.
   [[nodiscard]] constexpr size_type max_size() const noexcept
   {
@@ -121,6 +129,72 @@ private:
   friend class StlAllocator;
   MemoryAllocator* mBaseAllocator;
 };
+
+/// Stateless STL-compatible allocator backed by MemoryAllocator::GetDefault().
+///
+/// Use this adapter when container allocator state is undesirable and the
+/// process-wide default memory allocator is the intended backing resource.
+template <typename T>
+class DefaultStlAllocator
+{
+public:
+  using value_type = T;
+  using size_type = std::size_t;
+  using pointer = T*;
+  using const_pointer = const T*;
+  using is_always_equal = std::true_type;
+  using propagate_on_container_copy_assignment = std::true_type;
+  using propagate_on_container_move_assignment = std::true_type;
+  using propagate_on_container_swap = std::true_type;
+
+  template <typename U>
+  struct rebind
+  {
+    using other = DefaultStlAllocator<U>;
+  };
+
+  constexpr DefaultStlAllocator() noexcept = default;
+
+  constexpr DefaultStlAllocator(const DefaultStlAllocator& other) noexcept
+      = default;
+
+  template <typename U>
+  constexpr DefaultStlAllocator(const DefaultStlAllocator<U>& other) noexcept;
+
+  /// Allocates n * sizeof(T) bytes through MemoryAllocator::GetDefault().
+  [[nodiscard]] pointer allocate(size_type n, const void* hint = 0);
+
+  /// Deallocates storage previously returned by allocate().
+  void deallocate(pointer pointer, size_type n) noexcept;
+
+  /// Constructs an object at previously allocated storage.
+  template <typename U, typename... Args>
+  void construct(U* pointer, Args&&... args);
+
+  /// Destroys an object constructed by this allocator.
+  template <typename U>
+  void destroy(U* pointer) noexcept;
+
+  /// Upper bound on elements that can be allocated.
+  [[nodiscard]] constexpr size_type max_size() const noexcept
+  {
+    return std::numeric_limits<size_type>::max() / sizeof(T);
+  }
+};
+
+template <typename T, typename U>
+[[nodiscard]] constexpr bool operator==(
+    const DefaultStlAllocator<T>&, const DefaultStlAllocator<U>&) noexcept
+{
+  return true;
+}
+
+template <typename T, typename U>
+[[nodiscard]] constexpr bool operator!=(
+    const DefaultStlAllocator<T>&, const DefaultStlAllocator<U>&) noexcept
+{
+  return false;
+}
 
 } // namespace dart::common